+++ Assets/DataCenter/Safe/Scripts/Safe.cs
@@ -38,6 +38,10 @@ public class Safe : MonoBehaviour
         foreach (var gp in brokenGlass)
         {
             gp.gameObject.SetActive(true);
+            var tmpRb = gp.GetComponent<Rigidbody>();
+            tmpRb.AddTorque(new Vector3(20, 20, 50), ForceMode.Impulse);
+            var direction = Camera.main.transform.position - gp.position;
+            tmpRb.AddForce(direction.normalized * 3.2f, ForceMode.Impulse);
         }
 
     }