+++ ChangeLog.md
@@ -1,3 +1,11 @@
+## Develop
+* Added Shotgun safe
+* Added breakable glass
+* Added test Exit
+* Added floppy Sound
+* Added glass Sound
+* Added Shotgun sounds
+
 ## 0.2.0-alpha.19
 * Added Shotgun
 * Added Shotgun animation